TRUTH/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
